requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.3
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
gunicorn==20.1.0
MarkupSafe==2.1.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
